[EDIT] Packets attributes and methods. WIP

+++ Core/Entities/Packets/Models/DB/Announcements/S1S2Announcement/S1S2AnnouncementMethod.cs
@@ -12,7 +12,7 @@ public partial class S1S2Announcement : Announcement, IBaseEntity<S1S2Announceme
 
 	public S1S2Announcement(AnnouncementStruct adsStruct) : base(adsStruct)
 	{
-		SerialNumber = (int)adsStruct.SerialNumber;
+		SerialNumber = adsStruct.SerialNumber;
 		TrolleyNumber = (int)adsStruct.TrolleyNumber;
 	}
